# program/docs/api/openapi-college.json
{
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /api/v1/auth/login",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Unified API serving University, College, Secondary School, and Primary School management systems.",
    "title": "Education System API \u2014 College",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/complaints": {
      "post": {
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Complaints \u2014 Create Complaint",
        "tags": [
          "Complaints"
        ]
      }
    },
...
    "/api/complaints/{pk}": {
      "delete": {
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Complaints \u2014 Delete Complaint",
        "tags": [
          "Complaints"
        ]
      }
    },
    "/api/complaints/{pk}/escalate": {
      "post": {
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Complaints \u2014 Escalate",
        "tags": [
          "Complaints"
        ]
      }
    },
...
    "/api/first-aid": {
      "post": {
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "First-Aid \u2014 Report Incident",
        "tags": [
          "First-Aid"
        ]
      }
    },
    "/api/first-aid/{pk}": {
      "delete": {
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "First-Aid \u2014 Delete Incident",
        "tags": [
          "First-Aid"
        ]
      }
    },
...
    "/api/internal-verification/plans": {
      "post": {
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Internal-Verification \u2014 Create Plan",
        "tags": [
          "Internal-Verification"
        ]
      }
    },
    "/api/internal-verification/plans/{pk}": {
      "put": {
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Internal-Verification \u2014 Update Plan",
        "tags": [
          "Internal-Verification"
        ]
      }
    },
...
    "/api/tlevel/routes": {
      "post": {
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Tlevel \u2014 Create Route",
        "tags": [
          "Tlevel"
        ]
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "Current server",
      "url": "/"
    }
  ]
}